+++ cyberdrop_dl/ui/prompts/user_prompts.py
@@ -10,7 +10,7 @@ from cyberdrop_dl import __version__
 from cyberdrop_dl.ui.prompts import basic_prompts
 from cyberdrop_dl.ui.prompts.defaults import ALL_CHOICE, DONE_CHOICE, EXIT_CHOICE
 from cyberdrop_dl.utils.constants import BROWSERS, RESERVED_CONFIG_NAMES
-from cyberdrop_dl.utils.cookie_extraction import get_cookies_from_browsers
+from cyberdrop_dl.utils.cookie_management import get_cookies_from_browsers
 from cyberdrop_dl.utils.data_enums_classes.supported_domains import FORUMS, WEBSITES
 from cyberdrop_dl.utils.utilities import clear_term
